+++ src/main/kotlin/Puzzle.kt
@@ -2,7 +2,7 @@ interface Puzzle {
 
   val day: Int
 
-  fun solvePart1(input: String): Int
+  fun solvePart1(input: String): Long
 
-  fun solvePart2(input: String): Int
+  fun solvePart2(input: String): Long
 }
